Extract colors to ts
Extracts the color constants to a ts file
